finish moving
