--- .openapi-generator-ignore
@@ -1,34 +1,0 @@
-# OpenAPI Generator Ignore
-# Generated by openapi-generator https://github.com/openapitools/openapi-generator
-
-# Use this file to prevent files from being overwritten by the generator.
-# The patterns follow closely to .gitignore or .dockerignore.
-
-# As an example, the C# client generator defines ApiClient.cs.
-# You can make changes and tell OpenAPI Generator to ignore just this file by uncommenting the following line:
-#ApiClient.cs
-
-# You can match any string of characters against a directory, file or extension with a single asterisk (*):
-#foo/*/qux
-# The above matches foo/bar/qux and foo/baz/qux, but not foo/bar/baz/qux
-
-# You can recursively match patterns against a directory, file or extension with a double asterisk (**):
-#foo/**/qux
-# This matches foo/bar/qux, foo/baz/qux, and foo/bar/baz/qux
-
-# You can also negate patterns with an exclamation (!).
-# For example, you can ignore all files in a docs folder with the file extension .md:
-#docs/*.md
-# Then explicitly reverse the ignore rule for a single file:
-#!docs/README.md
-build.bat
-build.sh
-.sln
-Dockerfile
-Coflnet.SongVoter/Program.cs
-Coflnet.SongVoter/Startup.cs
-Coflnet.SongVoter/Dockerfile
-Coflnet.SongVoter/Coflnet.SongVoter.csproj
-Coflnet.SongVoter/Coflnet.SongVoter.sln
-Coflnet.SongVoter/appsettings.json
-Coflnet.SongVoter.sln
