--- Controllers/ListApi.cs
@@ -1,72 +1,0 @@
-/*
- * Songvoter
- *
- * Definition for songvoter API
- *
- * The version of the OpenAPI document: 0.0.1
- * Contact: support@coflnet.com
- * Generated by: https://openapi-generator.tech
- */
-
-using System;
-using System.Collections.Generic;
-using System.ComponentModel.DataAnnotations;
-using System.Threading.Tasks;
-using Microsoft.AspNetCore.Authorization;
-using Microsoft.AspNetCore.Mvc;
-using Microsoft.AspNetCore.Http;
-using Swashbuckle.AspNetCore.Annotations;
-using Swashbuckle.AspNetCore.SwaggerGen;
-using Newtonsoft.Json;
-using Coflnet.SongVoter.Attributes;
-using Coflnet.SongVoter.Models;
-
-namespace Coflnet.SongVoter.Controllers
-{ 
-    /// <summary>
-    /// 
-    /// </summary>
-    [ApiController]
-    public abstract class ListApiController : ControllerBase
-    { 
-        /// <summary>
-        /// Create a new playlist
-        /// </summary>
-        /// <param name="playList">An array of songIds to be added to the song</param>
-        /// <response code="200">successful created list</response>
-        [HttpPost]
-        [Route("/v1/lists")]
-        [Authorize]
-        [Consumes("application/json")]
-        [ValidateModelState]
-        [SwaggerOperation("CreatePlaylist")]
-        [SwaggerResponse(statusCode: 200, type: typeof(PlayList), description: "successful created list")]
-        public abstract Task<IActionResult> CreatePlaylist([FromBody]PlayList playList);
-
-        /// <summary>
-        /// Find playlist by ID
-        /// </summary>
-        /// <remarks>Returns a playList</remarks>
-        /// <param name="listId">ID of list to return</param>
-        /// <response code="200">successful operation</response>
-        [HttpGet]
-        [Route("/v1/lists/{listId}")]
-        [Authorize]
-        [ValidateModelState]
-        [SwaggerOperation("GetListById")]
-        [SwaggerResponse(statusCode: 200, type: typeof(PlayList), description: "successful operation")]
-        public abstract Task<IActionResult> GetListById([FromRoute (Name = "listId")][Required]string listId);
-
-        /// <summary>
-        /// Get playlist for active user
-        /// </summary>
-        /// <response code="200">successful response</response>
-        [HttpGet]
-        [Route("/v1/lists")]
-        [Authorize]
-        [ValidateModelState]
-        [SwaggerOperation("GetPlaylists")]
-        [SwaggerResponse(statusCode: 200, type: typeof(List<PlayList>), description: "successful response")]
-        public abstract Task<IActionResult> GetPlaylists();
-    }
-}
