+++ Properties/launchSettings.json
@@ -20,8 +20,8 @@
     "OpenAPI": {
       "commandName": "Project",
       "launchBrowser": true,
-      "launchUrl": "openapi",
-      "applicationUrl": "https://localhost:5001;http://localhost:5000",
+      "launchUrl": "api",
+      "applicationUrl": "http://localhost:5000",
       "environmentVariables": {
         "ASPNETCORE_ENVIRONMENT": "Development"
       }
