+++ SongVoter.csproj
@@ -1,8 +1,8 @@
 <Project Sdk="Microsoft.NET.Sdk.Web">
   <PropertyGroup>
-    <Description>A library generated from a OpenAPI doc</Description>
+    <Description>Songvoter.party main backend</Description>
     <Copyright>No Copyright</Copyright>
-    <Authors>OpenAPI</Authors>
+    <Authors>Coflnet</Authors>
     <TargetFramework>net6.0</TargetFramework>
     <GenerateDocumentationFile>true</GenerateDocumentationFile>
     <PreserveCompilationContext>true</PreserveCompilationContext>
